# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "UniAssist API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/tasks/{task_id}/grading-guidance": {
      "get": {
        "tags": [
          "Staff teaching tools"
        ],
        "summary": "Guidance",
        "operationId": "guidance_tasks__task_id__grading_guidance_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuidanceList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Staff teaching tools"
        ],
        "summary": "Save Guidance",
        "operationId": "save_guidance_tasks__task_id__grading_guidance_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuidanceCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuidanceInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/grading-guidance/{guidance_id}": {
      "get": {
        "tags": [
          "Staff teaching tools"
        ],
        "summary": "Guidance Version",
        "operationId": "guidance_version_tasks__task_id__grading_guidance__guidance_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          },
          {
            "name": "guidance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Guidance Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuidanceInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/analytics": {
      "get": {
        "tags": [
          "Staff teaching tools"
        ],
        "summary": "Analytics",
        "operationId": "analytics_tasks__task_id__analytics_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskAnalytics"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/analytics/reports": {
      "get": {
        "tags": [
          "Staff teaching tools"
        ],
        "summary": "Reports",
        "operationId": "reports_tasks__task_id__analytics_reports_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 30,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeachingReportList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Staff teaching tools"
        ],
        "summary": "Generate",
        "operationId": "generate_tasks__task_id__analytics_reports_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeachingReportInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommonIssue": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "severity": {
            "type": "string",
            "title": "Severity",
            "description": "info|warning|critical",
            "default": "warning"
          },
          "affected_estimate": {
            "type": "string",
            "title": "Affected Estimate",
            "description": "Human-readable estimate, e.g. 'many students' or '~40%'",
            "default": ""
          },
          "evidence": {
            "type": "string",
            "title": "Evidence",
            "description": "What in the cohort data supports this",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "title",
          "description"
        ],
        "title": "CommonIssue"
      },
...
      "CriterionStatistic": {
        "properties": {
          "criterion_id": {
            "type": "integer",
            "title": "Criterion Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "max_points": {
            "type": "number",
            "title": "Max Points"
          },
          "sample_count": {
            "type": "integer",
            "title": "Sample Count"
          },
          "average_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Score"
          },
          "low_score_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Low Score Count"
          }
        },
        "type": "object",
        "required": [
          "criterion_id",
          "name",
          "label",
          "max_points",
          "sample_count",
          "average_score",
          "low_score_count"
        ],
        "title": "CriterionStatistic"
      },
...
      "GuidanceCreate": {
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "content": {
            "type": "string",
            "maxLength": 12000,
            "title": "Content"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "content"
        ],
        "title": "GuidanceCreate"
      },
      "GuidanceInfo": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "task_id": {
            "type": "integer",
            "title": "Task Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "created_by": {
            "type": "integer",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "task_id",
          "version",
          "content",
          "created_by",
          "created_at"
        ],
        "title": "GuidanceInfo"
      },
      "GuidanceList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/GuidanceInfo"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "GuidanceList"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Misconception": {
        "properties": {
          "concept": {
            "type": "string",
            "title": "Concept"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "suggested_remediation": {
            "type": "string",
            "title": "Suggested Remediation",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "concept",
          "description"
        ],
        "title": "Misconception"
      },
...
      "ReportCreate": {
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "rubric_id": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Rubric Id"
          },
          "language": {
            "type": "string",
            "enum": [
              "en",
              "ar"
            ],
            "title": "Language",
            "default": "en"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "rubric_id"
        ],
        "title": "ReportCreate"
      },
...
      "RubricGroup": {
        "properties": {
          "rubric_id": {
            "type": "integer",
            "title": "Rubric Id"
          },
          "rubric_version": {
            "type": "integer",
            "title": "Rubric Version"
          },
          "student_count": {
            "type": "integer",
            "title": "Student Count"
          },
          "eligible": {
            "type": "boolean",
            "title": "Eligible"
          },
          "average_percentage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Percentage"
          },
          "minimum_percentage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minimum Percentage"
          },
          "maximum_percentage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maximum Percentage"
          },
          "below_half_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Below Half Count"
          },
          "criteria": {
            "items": {
              "$ref": "#/components/schemas/CriterionStatistic"
            },
            "type": "array",
            "title": "Criteria"
          },
          "severity_counts": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity Counts"
          },
          "mock_assessment_count": {
            "type": "integer",
            "title": "Mock Assessment Count"
          },
          "unknown_provenance_count": {
            "type": "integer",
            "title": "Unknown Provenance Count"
          }
        },
        "type": "object",
        "required": [
          "rubric_id",
          "rubric_version",
          "student_count",
          "eligible",
          "average_percentage",
          "minimum_percentage",
          "maximum_percentage",
          "below_half_count",
          "criteria",
          "severity_counts",
          "mock_assessment_count",
          "unknown_provenance_count"
        ],
        "title": "RubricGroup"
      },
...
      "TaskAnalytics": {
        "properties": {
          "task_id": {
            "type": "integer",
            "title": "Task Id"
          },
          "minimum_group_size": {
            "type": "integer",
            "title": "Minimum Group Size"
          },
          "released_count": {
            "type": "integer",
            "title": "Released Count"
          },
          "excluded_count": {
            "type": "integer",
            "title": "Excluded Count"
          },
          "input_fingerprint": {
            "type": "string",
            "title": "Input Fingerprint"
          },
          "groups": {
            "items": {
              "$ref": "#/components/schemas/RubricGroup"
            },
            "type": "array",
            "title": "Groups"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "minimum_group_size",
          "released_count",
          "excluded_count",
          "input_fingerprint",
          "groups"
        ],
        "title": "TaskAnalytics"
      },
...
      "TeachingReportInfo": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "task_id": {
            "type": "integer",
            "title": "Task Id"
          },
          "rubric_id": {
            "type": "integer",
            "title": "Rubric Id"
          },
          "rubric_version": {
            "type": "integer",
            "title": "Rubric Version"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "input_fingerprint": {
            "type": "string",
            "title": "Input Fingerprint"
          },
          "input_snapshot": {
            "$ref": "#/components/schemas/RubricGroup"
          },
          "result": {
            "$ref": "#/components/schemas/TeachingSuggestions"
          },
          "is_mock": {
            "type": "boolean",
            "title": "Is Mock"
          },
          "ai_metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TutorMetadata"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "stale": {
            "type": "boolean",
            "title": "Stale"
          }
        },
        "type": "object",
        "required": [
          "id",
          "task_id",
          "rubric_id",
          "rubric_version",
          "language",
          "input_fingerprint",
          "input_snapshot",
          "result",
          "is_mock",
          "ai_metadata",
          "created_at",
          "stale"
        ],
        "title": "TeachingReportInfo"
      },
      "TeachingReportList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TeachingReportInfo"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "TeachingReportList"
      },
      "TeachingSuggestions": {
        "properties": {
          "summary": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Summary"
          },
          "common_issues": {
            "items": {
              "$ref": "#/components/schemas/CommonIssue"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Common Issues"
          },
          "misconceptions": {
            "items": {
              "$ref": "#/components/schemas/Misconception"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Misconceptions"
          },
          "teaching_focus": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Teaching Focus"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "summary"
        ],
        "title": "TeachingSuggestions"
      },
...
      "TutorMetadata": {
        "properties": {
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "model_used": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Used"
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          },
          "ai_engine_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Engine Version"
          },
          "latency_ms": {
            "type": "number",
            "title": "Latency Ms"
          },
          "output_sanitized": {
            "type": "boolean",
            "title": "Output Sanitized",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "latency_ms"
        ],
        "title": "TutorMetadata"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}